requirements.txt
pandas
numpy
scikit-learn
scipy
torch
tqdm
matplotlib
PyYAML
fastapi
uvicorn
